Refactor(userAuth): 무분별한 setter 수정, dto경로를 명확하게 import
