+++ src/main/java/com/sparta/newsfeed/user/User.java
@@ -1,5 +1,6 @@
 package com.sparta.newsfeed.user;
 
+import com.sparta.newsfeed.profile.Profile;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,14 +8,13 @@ import lombok.Setter;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor
 @Table(name = "users")
 public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Long user_id;
 
     @Column(nullable = false, unique = true)
     private String username;
@@ -22,8 +22,16 @@ public class User {
     @Column(nullable = false)
     private String password;
 
+    @OneToOne
+    @JoinColumn(name = "profile_id")
+    private Profile profile;
+
     public User(String username, String password) {
         this.username = username;
+        this.password = password;
+    }
+
+    public void setPassword(String password) {
         this.password = password;
     }
 }
