+++ src/main/java/com/sparta/newsfeed/user/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.user.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Getter
-@Setter
 public class LoginRequestDto {
 
     @NotBlank
