+++ src/main/java/com/sparta/newsfeed/user/dto/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.user.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,7 +7,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Getter
-@Setter
 public class SignupRequestDto {
 
     @Pattern(regexp = "^[a-z0-9]{6,12}$")
